PKG -- [fcl] Move normalizers to a common normalizers folder
