+++ packages/fcl/src/current-user/exec-service/strategies/http-post.js
@@ -1,6 +1,6 @@
 import {fetchService} from "./utils/fetch-service"
-import {normalizePollingResponse} from "../../normalize/polling-response"
-import {normalizeLocalView} from "../../normalize/local-view"
+import {normalizePollingResponse} from "../../../normalizers/service/polling-response"
+import {normalizeLocalView} from "../../../normalizers/service/local-view"
 import {poll} from "./utils/poll"
 import {execLocal} from "../exec-local"
 import {VERSION} from "../../../VERSION"
