+++ packages/fcl/src/current-user/exec-service/strategies/iframe-rpc.js
@@ -1,6 +1,6 @@
 import {uid} from "@onflow/util-uid"
 import {frame} from "./utils/frame"
-import {normalizePollingResponse} from "../../normalize/polling-response"
+import {normalizePollingResponse} from "../../../normalizers/service/polling-response"
 import {VERSION} from "../../../VERSION"
 
 export function execIframeRPC(service, body, opts, config) {
