+++ packages/fcl/src/current-user/exec-service/strategies/pop-rpc.js
@@ -1,6 +1,6 @@
 import {uid} from "@onflow/util-uid"
 import {pop} from "./utils/pop"
-import {normalizePollingResponse} from "../../normalize/polling-response"
+import {normalizePollingResponse} from "../../../normalizers/service/polling-response"
 import {VERSION} from "../../../VERSION"
 
 export function execPopRPC(service, body, opts, config) {
