+++ packages/fcl/src/current-user/exec-service/strategies/tab-rpc.js
@@ -1,6 +1,6 @@
 import {uid} from "@onflow/util-uid"
 import {tab} from "./utils/tab"
-import {normalizePollingResponse} from "../../normalize/polling-response"
+import {normalizePollingResponse} from "../../../normalizers/service/polling-response"
 import {VERSION} from "../../../VERSION"
 
 export function execTabRPC(service, body, opts, config) {
