+++ packages/fcl/src/interaction-template-utils/derive-cadence-by-network.js
@@ -1,5 +1,5 @@
 import {invariant} from "@onflow/sdk"
-import {normalizeInteractionTemplate} from "../normalizers/normalize/interaction-template"
+import {normalizeInteractionTemplate} from "../normalizers/interaction-template/interaction-template"
 
 export function deriveCadenceByNetwork({network, template}) {
   invariant(
