+++ packages/fcl/src/interaction-template-utils/get-interaction-template-audits.js
@@ -2,7 +2,7 @@ import {config, invariant} from "@onflow/sdk"
 import {log, LEVELS} from "@onflow/util-logger"
 import {query} from "../exec/query.js"
 import {generateTemplateId} from "./generate-template-id.js"
-import {normalizeInteractionTemplate} from "./normalize/interaction-template.js"
+import {normalizeInteractionTemplate} from "../normalizers/interaction-template/interaction-template.js"
 
 export async function getInteractionTemplateAudits({template, auditors}) {
   invariant(
